+++ contact.php
@@ -26,25 +26,24 @@
             <h1>Contact Us</h1>
             
             <div class = "flex-container">
-                <div class = "flex-panel">
+                <div class = "flex-panel" style="margin:10px;">
                     <img class = "img-badge" src = "images/general/contact/business.png"></img>
                     <h5>Business Inquiries & Staff Applications</h5><br>
                     <p class="center"><a href="mailto:contact@sweetelitegame.com?Subject=Business%20Inquiry" target="_top">contact@sweetelitegame.com</a></p>
                 </div>
 
-                <div class = "flex-panel">
+                <div class = "flex-panel" style="margin:10px;">
                     <img class = "img-badge" src = "images/general/contact/help.png"></img>
                     <h5>Player Support & Questions</h5><br><br>
                     <p class="center"><a href="mailto:support@sweetelitegame.com?Subject=Business%20Inquiry" target="_top">support@sweetelitegame.com</a></p>
                 </div>
 
-                <div class = "flex-panel">
+                <div class = "flex-panel" style="margin:10px;">
                     <img class = "img-badge" src = "images/general/contact/love.png"></img>
                     <h5>General Love</h5><br><br>
                     <p class="center"><a href="mailto:sweetelitestaff@gmail.com?Subject=Business%20Inquiry" target="_top">sweetelitestaff@gmail.com</a></p>
                 </div>
             </div>
-
 		</div>
 	</div>